Add basic process

+++ script.sh
@@ -1,21 +1,29 @@
 #! /bin/zsh
 
-cd lb01
-vagrant destroy -f
-vagrant up
-cd ..
+if [ "$1" = "all" ] || [ "$1" = "lb01" ]; then
+  cd lb01
+  vagrant destroy -f
+  vagrant up
+  cd ..
+fi
 
-cd app01
-vagrant destroy -f
-vagrant up
-cd ..
+if [ "$1" = "all" ] || [ "$1" = "app01" ]; then
+  cd app01
+  vagrant destroy -f
+  vagrant up
+  cd ..
+fi
 
-cd app02
-vagrant destroy -f
-vagrant up
+if [ "$1" = "all" ] || [ "$1" = "app02" ]; then
+  cd app02
+  vagrant destroy -f
+  vagrant up
 cd ..
+fi
 
-cd db01
-vagrant destroy -f
-vagrant up
-cd ..
+if [ "$1" = "all" ] || [ "$1" = "db01" ]; then
+  cd db01
+  vagrant destroy -f
+  vagrant up
+  cd ..
+fi
